+++ gamedata/__init__.py
@@ -1,7 +1,7 @@
 from .background import Background
 from .feat import Feat
 from .item import Item
-from .klass import Class
+from .klass import Class, Subclass
 from .monster import Monster
 from .race import Race
 from .spell import Spell
